agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kube-rca-agent"
version = "0.1.0"
description = "Kube-RCA agent service"
readme = "README.md"
requires-python = ">=3.10"  # Strands Agents supports Python 3.10+.
dependencies = [
  "fastapi>=0.115.12,<1.0.0",
  "pydantic>=2.4.0,<3.0.0",
  "strands-agents[gemini]>=1.20.0,<2.0.0",
  "kubernetes>=34.0.0,<35.0.0",  # Latest checked: 34.1.0.
  "uvicorn[standard]>=0.34.2,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "ruff>=0.13.0,<0.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
